Extract config to it's own file
