+++ kafka/client.go
@@ -1,11 +1,8 @@
 package kafka
 
 import (
-	"crypto/tls"
-	"crypto/x509"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"time"
 
@@ -23,18 +20,6 @@ type Client struct {
 	kafkaConfig   *sarama.Config
 	config        *Config
 	supportedAPIs map[int]int
-}
-
-type Config struct {
-	BootstrapServers *[]string
-	Timeout          int
-	CACertFile       string
-	ClientCertFile   string
-	ClientCertKey    string
-	TLSEnabled       bool
-	SkipTLSVerify    bool
-	SASLUsername     string
-	SASLPassword     string
 }
 
 func NewClient(config *Config) (*Client, error) {
@@ -96,7 +81,6 @@ func (c *Client) populateAPIVersions() {
 
 func (c *Client) DeleteTopic(t string) error {
 	broker, err := c.client.Controller()
-
 	if err != nil {
 		return err
 	}
@@ -126,7 +110,6 @@ func (c *Client) DeleteTopic(t string) error {
 
 func (c *Client) UpdateTopic(topic Topic) error {
 	broker, err := c.client.Controller()
-
 	if err != nil {
 		return err
 	}
@@ -155,9 +138,7 @@ func (c *Client) UpdateTopic(topic Topic) error {
 
 func (c *Client) CreateTopic(t Topic) error {
 	broker, err := c.client.Controller()
-
 	if err != nil {
-		log.Printf("[WARN] Could get an available broker %s", err)
 		return err
 	}
 
@@ -189,11 +170,10 @@ func (c *Client) CreateTopic(t Topic) error {
 
 func (c *Client) AddPartitions(t Topic) error {
 	broker, err := c.client.Controller()
-
 	if err != nil {
-		log.Printf("[WARN] DERP %s", err)
 		return err
 	}
+
 	timeout := time.Duration(c.config.Timeout) * time.Second
 	log.Printf("[DEBUG] b of size %d", 1)
 	tp := map[string]*sarama.TopicPartition{
@@ -201,6 +181,7 @@ func (c *Client) AddPartitions(t Topic) error {
 			Count: t.Partitions,
 		},
 	}
+
 	log.Printf("[DEBUG] b of size %d", 2)
 	req := &sarama.CreatePartitionsRequest{
 		TopicPartitions: tp,
@@ -338,66 +319,4 @@ func (c *Client) availableBroker() (*sarama.Broker, error) {
 	}
 
 	return nil, fmt.Errorf("No Available Brokers @ %v", brokers)
-}
-
-func (c *Config) newKafkaConfig() (*sarama.Config, error) {
-	kafkaConfig := sarama.NewConfig()
-	kafkaConfig.Version = sarama.V1_0_0_0
-
-	if c.saslEnabled() {
-		kafkaConfig.Net.SASL.Enable = true
-		kafkaConfig.Net.SASL.Password = c.SASLPassword
-		kafkaConfig.Net.SASL.User = c.SASLUsername
-	}
-
-	if c.TLSEnabled {
-		tlsConfig, err := newTLSConfig(
-			c.ClientCertFile,
-			c.ClientCertKey,
-			c.CACertFile)
-
-		if err != nil {
-			return kafkaConfig, err
-		}
-
-		kafkaConfig.Net.TLS.Enable = true
-		kafkaConfig.Net.TLS.Config = tlsConfig
-		kafkaConfig.Net.TLS.Config.InsecureSkipVerify = c.SkipTLSVerify
-	}
-
-	return kafkaConfig, nil
-}
-
-func (c *Config) saslEnabled() bool {
-	return c.SASLUsername != "" || c.SASLPassword != ""
-}
-
-func newTLSConfig(clientCertFile, clientKeyFile, caCertFile string) (*tls.Config, error) {
-	tlsConfig := tls.Config{}
-
-	// Load client cert
-	if clientCertFile != "" && clientKeyFile != "" {
-		cert, err := tls.LoadX509KeyPair(clientCertFile, clientKeyFile)
-		if err != nil {
-			return &tlsConfig, err
-		}
-		tlsConfig.Certificates = []tls.Certificate{cert}
-	} else {
-		log.Println("[WARN] skipping TLS client config")
-	}
-
-	if caCertFile == "" {
-		log.Println("[WARN] no CA file set skipping")
-		return &tlsConfig, nil
-	}
-	// Load CA cert
-	caCert, err := ioutil.ReadFile(caCertFile)
-	if err != nil {
-		return &tlsConfig, err
-	}
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
-	tlsConfig.RootCAs = caCertPool
-	tlsConfig.BuildNameToCertificate()
-	return &tlsConfig, err
 }
