+++ kafka/config.go
@@ -1,0 +1,84 @@
+package kafka
+
+import (
+	"crypto/tls"
+	"crypto/x509"
+	"io/ioutil"
+	"log"
+
+	"github.com/Shopify/sarama"
+)
+
+type Config struct {
+	BootstrapServers *[]string
+	Timeout          int
+	CACertFile       string
+	ClientCertFile   string
+	ClientCertKey    string
+	TLSEnabled       bool
+	SkipTLSVerify    bool
+	SASLUsername     string
+	SASLPassword     string
+}
+
+func (c *Config) newKafkaConfig() (*sarama.Config, error) {
+	kafkaConfig := sarama.NewConfig()
+	kafkaConfig.Version = sarama.V1_0_0_0
+
+	if c.saslEnabled() {
+		kafkaConfig.Net.SASL.Enable = true
+		kafkaConfig.Net.SASL.Password = c.SASLPassword
+		kafkaConfig.Net.SASL.User = c.SASLUsername
+	}
+
+	if c.TLSEnabled {
+		tlsConfig, err := newTLSConfig(
+			c.ClientCertFile,
+			c.ClientCertKey,
+			c.CACertFile)
+
+		if err != nil {
+			return kafkaConfig, err
+		}
+
+		kafkaConfig.Net.TLS.Enable = true
+		kafkaConfig.Net.TLS.Config = tlsConfig
+		kafkaConfig.Net.TLS.Config.InsecureSkipVerify = c.SkipTLSVerify
+	}
+
+	return kafkaConfig, nil
+}
+
+func (c *Config) saslEnabled() bool {
+	return c.SASLUsername != "" || c.SASLPassword != ""
+}
+
+func newTLSConfig(clientCertFile, clientKeyFile, caCertFile string) (*tls.Config, error) {
+	tlsConfig := tls.Config{}
+
+	// Load client cert
+	if clientCertFile != "" && clientKeyFile != "" {
+		cert, err := tls.LoadX509KeyPair(clientCertFile, clientKeyFile)
+		if err != nil {
+			return &tlsConfig, err
+		}
+		tlsConfig.Certificates = []tls.Certificate{cert}
+	} else {
+		log.Println("[WARN] skipping TLS client config")
+	}
+
+	if caCertFile == "" {
+		log.Println("[WARN] no CA file set skipping")
+		return &tlsConfig, nil
+	}
+	// Load CA cert
+	caCert, err := ioutil.ReadFile(caCertFile)
+	if err != nil {
+		return &tlsConfig, err
+	}
+	caCertPool := x509.NewCertPool()
+	caCertPool.AppendCertsFromPEM(caCert)
+	tlsConfig.RootCAs = caCertPool
+	tlsConfig.BuildNameToCertificate()
+	return &tlsConfig, err
+}
